KJS: fix index truncation
Everyone calling this function uses a size_t as variable and we have a size_t
internally as index variable, too. So no need to needlessly truncate this and
cast it to a signed int.

+++ kjs/collector.cpp
@@ -101,7 +101,7 @@ struct BlockList {
     size_t m_used;
     size_t m_capacity;
 
-    CollectorBlock* operator[](int pos) {
+    CollectorBlock* operator[](size_t pos) {
         return m_data[pos];
     }
